+++ controllers/auth/getCurrent/getCurrent.js
@@ -1,0 +1,8 @@
+const getCurrent = async (req, res) => {
+    const { email, subscription } = req.body;
+    res.status(200).json({
+      email,
+      subscription,
+    });
+  };
+  module.exports = getCurrent;
