+++ controllers/auth/logout/logout.js
@@ -1,0 +1,11 @@
+const { User } = require("../../../models/user");
+
+const logout = async (req, res) => {
+    const { _id } = req.user;
+    await User.findByIdAndUpdate(_id, { token: "" });
+    res.status(204).json({
+      message: "No Content",
+    });
+  };
+
+  module.exports = logout;
