+++ controllers/diary/index.js
@@ -1,0 +1,7 @@
+// const { ctrlWrapper } = require("../../helpers");
+
+// const register = require("./register/register");
+
+module.exports = {
+	// register: ctrlWrapper(register),
+};
