+++ routes/api/training.js
@@ -1,0 +1,8 @@
+const express = require("express");
+// const ctrl = require("../../controllers/diary/index");
+const router = express.Router();
+// const { schemas } = require("../../models/training");
+
+// router.get(... , ctrl.register)
+
+module.exports = router;
